* 'master' of github.com:vispy/vispy:
  Changed texture API a bit, making separate method for setting subdata. The reason is that setting data and subdata should behave quite differently (as it does now). set_data and set_storage resets the image, so for multiple calls, the last override earlier calls. For subdata, however, you want to take every call, because the texture might be updated in blocks.
  Removed push_enable and pop_enable. They are not used. We may want *something* like this later, as cyrille just proposed, but probably not in this form.
  Improving docs. The example building and gl api building are done in a sphinx extension. The above buildings also clean up after themselves. Added the vispy.gl.ext namespace. Added some notes about current limitations of vispy.oogl.
  Added type check for matrix uniforms

+++ doc/examplesgenerator.py
@@ -9,16 +9,26 @@ THISDIR = os.path.dirname(os.path.abspath(__file__))
 EXAMPLESDIR = os.path.join(THISDIR, '..', 'examples')
 OUTPUTDIR = os.path.join(THISDIR, 'examples')
 
+
+def clean():
+    # Clear examples dir
+    if os.path.isdir(OUTPUTDIR):
+        for fname in os.listdir(OUTPUTDIR):
+            if fname.endswith('.rst'):
+                os.remove(os.path.join(OUTPUTDIR, fname))
+        os.rmdir(OUTPUTDIR)
+    # Clean examples file
+    fname = os.path.join(THISDIR, 'examples.rst')
+    if os.path.isfile(fname):
+        os.remove(fname)
+
+
 def main():
-        
+    
+    clean()
+    
     if not os.path.isdir(OUTPUTDIR):
         os.mkdir(OUTPUTDIR)
-    
-    # Clear examples dir
-    for fname in os.listdir(OUTPUTDIR):
-        if fname.endswith('.rst'):
-            os.remove(os.path.join(OUTPUTDIR, fname))
-    
     
     example_names = []
